==> notebooks/1. Eliminación columnas.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
from sklearn.impute import SimpleImputer
from sklearn import linear_model 
from sklearn.model_selection import train_test_split
from sklearn import metrics
from sklearn import preprocessing
from sklearn.model_selection import GridSearchCV
from matplotlib import pyplot as plt
from sklearn.metrics import make_scorer
from sklearn.metrics import accuracy_score
from sklearn import tree
from sklearn.preprocessing import scale, StandardScaler, OneHotEncoder
import sklearn
from sklearn.decomposition import PCA
from sklearn.impute import KNNImputer

df = pd.read_excel("esg_project21.xlsx")
df.head(5)

### Reemplazo NaNs, eliminación y tratmiento de las columnas (basicos)

def preproces(df):
    
    # Limpieza de nombres de columnas
    df.columns = df.columns.str.strip()
    
    # Reemplazo los n.d. y valores vacíos por NaN 
    df.replace({"n.d.": np.nan, "": np.nan, 'n.s.': np.nan}, inplace=True)
    
    # Calculo el porcentaje de NaN por columna
    nan_per = df.isna().mean() * 100
    
    # Filtro las columnas con mas del 80% de NaN
    columnas_mas_80 = nan_per[nan_per > 80].index
    
    # Elimino las columnas con más del 80% de NaN
    df.drop(columns=columnas_mas_80, inplace=True)
    
    return df

# Supongamos que 'df' es tu DataFrame original y 'preproces' es tu función
clening_basic_df = preproces(df)

# Guardar el DataFrame en un archivo CSV
clening_basic_df.to_csv('clening_basic_data.csv', index=False)

print("Archivo guardado como 'clening_basic_data.csv'")
